+++ test/gulp.js
@@ -1,128 +1,99 @@
-'use strict';
+import '../src'
+import bufferEqual from 'buffer-equal'
+import convert from 'convert-source-map'
+import gulp from 'gulp'
+import timeout from 'timeout-then'
+import rump from 'rump'
+import {colors} from 'gulp-util'
+import {readFile, writeFile} from 'mz/fs'
+import {resolve, sep} from 'path'
+import {spy} from 'sinon'
 
-// Temporary fix until old LoDash is updated in some Gulp dependency
-Object.getPrototypeOf.toString = function() {
-  return 'function getPrototypeOf() { [native code] }';
-};
+const {stripColor} = colors,
+      protocol = process.platform === 'win32' ? 'file:///' : 'file://'
 
-var assert = require('assert');
-var bufferEqual = require('buffer-equal');
-var co = require('co');
-var convert = require('convert-source-map');
-var fs = require('mz/fs');
-var gulp = require('gulp');
-var util = require('gulp-util');
-var path = require('path');
-var sinon = require('sinon');
-var sleep = require('timeout-then');
-var rump = require('../lib');
-var protocol = process.platform === 'win32' ? 'file:///' : 'file://';
+describe('tasks', function() {
+  this.timeout(0)
 
-describe('rump scripts tasks', function() {
-  beforeEach(function() {
+  beforeEach(() => {
     rump.configure({
       environment: 'development',
       paths: {
-        source: {
-          root: 'test/src/js',
-          scripts: ''
-        },
-        destination: {
-          root: 'tmp',
-          scripts: ''
-        }
-      }
-    });
-  });
+        source: {root: 'test/src', scripts: ''},
+        destination: {root: 'tmp', scripts: ''},
+      },
+    })
+  })
 
-  it('are added and defined', function() {
-    var callback = sinon.spy();
-    rump.on('gulp:main', callback);
-    rump.on('gulp:scripts', callback);
-    rump.addGulpTasks({prefix: 'spec'});
-    // TODO Remove no callback check on next major core update
-    assert(!callback.called || callback.calledTwice);
-    assert(gulp.tasks['spec:info:scripts']);
-    assert(gulp.tasks['spec:build:scripts']);
-  });
+  it('are added and defined', () => {
+    const callback = spy()
+    rump.on('gulp:main', callback)
+    rump.on('gulp:scripts', callback)
+    rump.addGulpTasks({prefix: 'spec'})
+    callback.should.be.calledTwice()
+    gulp.tasks['spec:info:scripts'].should.be.ok()
+    gulp.tasks['spec:build:scripts'].should.be.ok()
+  })
 
-  it('displays correct information in info task', function() {
-    var oldLog = console.log;
-    var logs = [];
-    console.log = function() {
-      logs.push(util.colors.stripColor(Array.from(arguments).join(' ')));
-    };
-    gulp.start('spec:info');
-    console.log = oldLog;
-    assert(logs.some(hasPaths));
-    assert(logs.some(hasJsFile));
-    assert(!logs.some(hasLibFile));
-  });
+  it('displays correct information in info task', () => {
+    const logs = [],
+          {log} = console
 
-  describe('for building', function() {
-    var originals;
+    console.log = (...args) => logs.push(stripColor(args.join(' ')))
+    gulp.start('spec:info')
+    console.log = log
+    logs.slice(-6).should.eql([
+      '',
+      '--- Scripts v0.7.0',
+      `Processed scripts from test${sep}src are copied with source maps to tmp`,
+      'Affected files:',
+      'index.js',
+      '',
+    ])
+  })
 
-    before(co.wrap(function*() {
-      originals = yield [
-        fs.readFile('test/src/js/index.js'),
-        fs.readFile('test/src/js/lib/index.js')
-      ];
-    }));
+  describe('for building', () => {
+    let originals
 
-    before(function(done) {
-      gulp.task('postbuild', ['spec:watch'], function() {
-        done();
-      });
-      gulp.start('postbuild');
-    });
+    before(async(done) => {
+      originals = await Promise.all([
+        readFile('test/src/index.js'),
+        readFile('test/src/lib/index.js'),
+      ])
+      gulp.task('postbuild', ['spec:watch'], () => done())
+      gulp.start('postbuild')
+    })
 
-    afterEach(co.wrap(function*() {
-      yield sleep(800);
-      yield [
-        fs.writeFile('test/src/js/index.js', originals[0]),
-        fs.writeFile('test/src/js/lib/index.js', originals[1])
-      ];
-      yield sleep(800);
-    }));
+    afterEach(async() => {
+      await timeout(1000)
+      await Promise.all([
+        writeFile('test/src/index.js', originals[0]),
+        writeFile('test/src/lib/index.js', originals[1]),
+      ])
+      await timeout(1000)
+    })
 
-    it('handles updates', co.wrap(function*() {
-      var firstContent = yield fs.readFile('tmp/index.js');
-      yield sleep(800);
-      fs.writeFileSync('test/src/js/lib/index.js', 'module.exports = "";');
-      yield sleep(800);
-      var secondContent = yield fs.readFile('tmp/index.js');
-      assert(!bufferEqual(firstContent, secondContent));
-    }));
+    it('handles updates', async() => {
+      const firstContent = await readFile('tmp/index.js')
+      let secondContent
+      await timeout(1000)
+      await writeFile('test/src/lib/index.js', 'module.exports = "";')
+      await timeout(1000)
+      secondContent = await readFile('tmp/index.js')
+      bufferEqual(firstContent, secondContent).should.be.false()
+    })
 
-    it('handles source maps in development', co.wrap(function*() {
-      var content = yield fs.readFile('tmp/index.js');
-      var sourceMap = convert.fromSource(content.toString());
-      var exists = yield sourceMap
-            .getProperty('sources')
-            .slice(1)
-            .filter(identity)
-            .map(checkIfExists);
-      exists.forEach(assert);
-    }));
-  });
-});
-
-function hasJsFile(log) {
-  return log === 'index.js';
-}
-
-function hasLibFile(log) {
-  return log.includes(path.join('lib', 'index.js'));
-}
-
-function hasPaths(log) {
-  return log.includes(path.join('test', 'src', 'js')) && log.includes('tmp');
-}
-
-function identity(x) {
-  return x;
-}
-
-function checkIfExists(url) {
-  return fs.exists(url.replace(protocol, '').split('/').join(path.sep));
-}
+    it('handles source maps in development', async() => {
+      const content = await readFile('tmp/index.js'),
+            paths = convert
+              .fromSource(content.toString())
+              .getProperty('sources')
+              .sort()
+              .filter(x => x)
+              .map(x => x.replace(protocol, '').split('/').join(sep))
+      paths.should.eql([
+        resolve('test/src/index.js'),
+      ])
+    })
+  })
+})